Rename component lifecycle and hooks callback functions
