+++ packages/core/docs/functions/getFunctionComponentId.md
@@ -4,7 +4,6 @@
 
 ```ts
 function getFunctionComponentId(context: RuleContext, node: TSESTreeFunction): 
-  | JSXElement
   | ArrayExpression
   | ArrayPattern
   | ArrowFunctionExpression
@@ -19,6 +18,7 @@ function getFunctionComponentId(context: RuleContext, node: TSESTreeFunction):
   | FunctionExpression
   | Identifier
   | ImportExpression
+  | JSXElement
   | JSXFragment
   | BigIntLiteral
   | BooleanLiteral
@@ -65,7 +65,6 @@ function getFunctionComponentId(context: RuleContext, node: TSESTreeFunction):
 
 ## Returns
 
-  \| `JSXElement`
   \| `ArrayExpression`
   \| `ArrayPattern`
   \| `ArrowFunctionExpression`
@@ -80,6 +79,7 @@ function getFunctionComponentId(context: RuleContext, node: TSESTreeFunction):
   \| `FunctionExpression`
   \| `Identifier`
   \| `ImportExpression`
+  \| `JSXElement`
   \| `JSXFragment`
   \| `BigIntLiteral`
   \| `BooleanLiteral`
