+++ packages/core/src/component/component-method-callback.ts
@@ -1,14 +1,14 @@
 import * as AST from "@eslint-react/ast";
 import type { TSESTree } from "@typescript-eslint/types";
 
-import { isComponentDidMount, isComponentWillUnmount } from "./component-method";
+import { isComponentDidMount, isComponentWillUnmount } from "./component-method-is";
 
 /**
  * Checks if the node is a function of `componentDidMount`
  * @param node The AST node to check
  * @returns `true` if the node is a function of `componentDidMount`
  */
-export function isFunctionOfComponentDidMount(node: TSESTree.Node) {
+export function isComponentDidMountCallback(node: TSESTree.Node) {
   return AST.isFunction(node)
     && isComponentDidMount(node.parent)
     && node.parent.value === node;
@@ -19,7 +19,7 @@ export function isFunctionOfComponentDidMount(node: TSESTree.Node) {
  * @param node The AST node to check
  * @returns `true` if the node is a function of `componentWillUnmount`
  */
-export function isFunctionOfComponentWillUnmount(node: TSESTree.Node) {
+export function isComponentWillUnmountCallback(node: TSESTree.Node) {
   return AST.isFunction(node)
     && isComponentWillUnmount(node.parent)
     && node.parent.value === node;
