+++ packages/core/src/component/component-method-is.ts
