+++ packages/core/src/component/component-phase-helpers.ts
@@ -2,8 +2,8 @@ import type * as AST from "@eslint-react/ast";
 import { dual } from "@eslint-react/eff";
 import { match } from "ts-pattern";
 
-import { isFunctionOfUseEffectCleanup, isFunctionOfUseEffectSetup } from "../hook/hook-parts";
-import { isFunctionOfComponentDidMount, isFunctionOfComponentWillUnmount } from "./component-method-parts";
+import { isUseEffectCleanupCallback, isUseEffectSetupCallback } from "../hook/hook-callback";
+import { isComponentDidMountCallback, isComponentWillUnmountCallback } from "./component-method-callback";
 import { type ComponentPhaseKind, ComponentPhaseRelevance } from "./component-phase";
 
 export const isInversePhase: {
@@ -13,9 +13,9 @@ export const isInversePhase: {
 
 export function getPhaseKindOfFunction(node: AST.TSESTreeFunction) {
   return match<AST.TSESTreeFunction, ComponentPhaseKind | null>(node)
-    .when(isFunctionOfUseEffectSetup, () => "setup")
-    .when(isFunctionOfUseEffectCleanup, () => "cleanup")
-    .when(isFunctionOfComponentDidMount, () => "mount")
-    .when(isFunctionOfComponentWillUnmount, () => "unmount")
+    .when(isUseEffectSetupCallback, () => "setup")
+    .when(isUseEffectCleanupCallback, () => "cleanup")
+    .when(isComponentDidMountCallback, () => "mount")
+    .when(isComponentWillUnmountCallback, () => "unmount")
     .otherwise(() => null);
 }
