+++ packages/core/src/component/index.ts
@@ -7,8 +7,8 @@ export * from "./component-id";
 export * from "./component-init-path";
 export * from "./component-is";
 export type * from "./component-kind";
-export * from "./component-method";
-export * from "./component-method-parts";
+export * from "./component-method-callback";
+export * from "./component-method-is";
 export * from "./component-name";
 export * from "./component-phase";
 export * from "./component-phase-helpers";
