+++ packages/core/src/hook/index.ts
@@ -1,6 +1,6 @@
+export * from "./hook-callback";
 export * from "./hook-collector";
 export * from "./hook-id";
 export * from "./hook-is";
 export * from "./hook-name";
-export * from "./hook-parts";
 export type * from "./hook-semantic-node";
